Partial code for replaygain.
It does not write replaygain, but it can compute it correctly. I'm not
sure how I'm going to get the numbers into the files in an efficient
manner, and I still need to compute the peaks.

Well it still does write it using metaflac, but that has the
disadvantage of rewriting every file.