Bump up to C++17
Compilation was broken so I figured might as well bump up to C++ 17.

This also fixes a double-close of the input file. The file is closed by
libflac when the Decoder goes out of scope. It's also closed when the
File_handle goes out of scope. The fix was to release the FILE pointer
after its ownership transfers to the decoder.

I guess I'm not sure if libsox also runs into this :(.

* Replace scoped_array/scoped_ptr/auto_ptr with unique_ptr
* Replaced boost shared_ptr with std
* Drop throw() specifiers. Replaced with doxygen-style comments
* Encoder/Decoder were missing virtual destructors. Encoder also forgot
  to subclass Basic_encoder
* Replaced some asserts with runtime exceptions. Which in retrospect
  could have just been an include of <cassert>
* Also switch to C11 for no reason